+++ lib/graphql/execution/interpreter.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require "graphql/execution/interpreter/argument_value"
 require "graphql/execution/interpreter/arguments"
 require "graphql/execution/interpreter/arguments_cache"
 require "graphql/execution/interpreter/execution_errors"
